refactor(VERASettings): extract ActiveCallWarningText, ActiveCallFooter reusable views
- Add `ActiveCallWarningText` view to replace repeated "Cannot be changed during an active call" text blocks across settings sections
- Add `ActiveCallFooter` view to encapsulate the isInActiveCall ? warning : description footer pattern
- Replace all 16 raw occurrences across AudioSectionView, VideoSectionView, GeneralSectionView, and StatisticsSectionScreen
- Add spring animation to chevron rotation in ExpandableParticipantHeaderRow
