+++ VERA/VERASettings/VERASettings/UI/View/Sections/SettingsDivider.swift
@@ -1,0 +1,14 @@
+//
+//  Created by Vonage on 02/07/2026.
+//
+
+import SwiftUI
+import VERACommonUI
+
+struct SettingsDivider: View {
+    var body: some View {
+        Rectangle()
+            .fill(VERACommonUIAsset.SemanticColors.border.swiftUIColor)
+            .frame(height: 1)
+    }
+}
